modal for confirming overwrite
